+++ gateway/batchType_easyjson_test.go
@@ -4,6 +4,7 @@ package gateway
 
 import (
 	json "encoding/json"
+
 	easyjson "github.com/mailru/easyjson"
 	jlexer "github.com/mailru/easyjson/jlexer"
 	jwriter "github.com/mailru/easyjson/jwriter"
